💄  Ajuste de mosodo de mostrar los componentes
