+++ src/app/components/articles/article/article.component.scss
@@ -33,14 +33,6 @@
             padding: 0;
         }
     }
-
-    // &__redirect {
-    //     display: flex;
-    //     flex-direction: row;
-    //     justify-content: center;
-    //     align-items: center;
-    //     gap: 1rem;
-    // }
 }
