+++ src/app/components/projects/projects.module.ts
@@ -1,11 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProjectsComponent } from './projects/projects.component';
-import { SharedModule } from '@shared/shared.module';
 
 @NgModule({
   declarations: [ProjectsComponent],
-  imports: [CommonModule, SharedModule],
+  imports: [CommonModule],
   exports: [ProjectsComponent],
 })
 export class ProjectsModule {}
