+++ src/app/components/projects/projects/projects.component.html
@@ -1,37 +1,35 @@
-<section class="projects">
-  <article class="projects__cards" *ngFor="let project of projects|randomlist:3 ">
-    <div class="projects__cards-info">
-      <h2>{{ project?.name  }}</h2>
-      <span>{{project?.updated_at | date }}</span>
-      <p>
-        {{ project?.description }}
-      </p>
-      <span *ngFor="let topic of project?.topics">
-        {{topic}}
-      </span>
+<article class="cards">
+  <div class="cards__info">
+    <h2>{{ project?.name  }}</h2>
+    <span>{{project?.updated_at | date }}</span>
+    <p>
+      {{ project?.description }}
+    </p>
+    <span *ngFor="let topic of project?.topics">
+      {{topic}}
+    </span>
 
-    </div>
-    <div class="projects__cards-actions">
-      <svg (click)="share(project.html_url)" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img"
-        width="1.09em" height="1em" preserveAspectRatio="xMidYMid meet" viewBox="0 0 1664 1536" width="1.5em"
-        height="1.5em">
-        <path fill="currentColor"
-          d="M1408 989v259q0 119-84.5 203.5T1120 1536H288q-119 0-203.5-84.5T0 1248V416q0-119 84.5-203.5T288 128h255q13 0 22.5 9.5T575 160q0 27-26 32q-77 26-133 60q-10 4-16 4H288q-66 0-113 47t-47 113v832q0 66 47 113t113 47h832q66 0 113-47t47-113v-214q0-19 18-29q28-13 54-37q16-16 35-8q21 9 21 29zm237-496l-384 384q-18 19-45 19q-12 0-25-5q-39-17-39-59V640H992q-323 0-438 131q-119 137-74 473q3 23-20 34q-8 2-12 2q-16 0-26-13q-10-14-21-31t-39.5-68.5T312 1068t-38.5-114T256 832q0-49 3.5-91t14-90t28-88t47-81.5t68.5-74t94.5-61.5T636 297.5T795.5 267T992 256h160V64q0-42 39-59q13-5 25-5q26 0 45 19l384 384q19 19 19 45t-19 45z" />
+  </div>
+  <div class="cards__actions">
+    <svg (click)="share(project.html_url)" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img"
+      width="1.09em" height="1em" preserveAspectRatio="xMidYMid meet" viewBox="0 0 1664 1536" width="1.5em"
+      height="1.5em">
+      <path fill="currentColor"
+        d="M1408 989v259q0 119-84.5 203.5T1120 1536H288q-119 0-203.5-84.5T0 1248V416q0-119 84.5-203.5T288 128h255q13 0 22.5 9.5T575 160q0 27-26 32q-77 26-133 60q-10 4-16 4H288q-66 0-113 47t-47 113v832q0 66 47 113t113 47h832q66 0 113-47t47-113v-214q0-19 18-29q28-13 54-37q16-16 35-8q21 9 21 29zm237-496l-384 384q-18 19-45 19q-12 0-25-5q-39-17-39-59V640H992q-323 0-438 131q-119 137-74 473q3 23-20 34q-8 2-12 2q-16 0-26-13q-10-14-21-31t-39.5-68.5T312 1068t-38.5-114T256 832q0-49 3.5-91t14-90t28-88t47-81.5t68.5-74t94.5-61.5T636 297.5T795.5 267T992 256h160V64q0-42 39-59q13-5 25-5q26 0 45 19l384 384q19 19 19 45t-19 45z" />
+    </svg>
+    <a [href]="project?.html_url" target="_blank">
+      <svg xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img" width="1em" height="1em"
+        preserveAspectRatio="xMidYMid meet" viewBox="0 0 48 48" width="1.5em" height="1.5em">
+        <path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
+          d="M24 2.5a21.5 21.5 0 0 0-6.8 41.9c1.08.2 1.47-.46 1.47-1v-3.65c-6 1.3-7.24-2.88-7.24-2.88A5.7 5.7 0 0 0 9 33.68c-1.95-1.33.15-1.31.15-1.31a4.52 4.52 0 0 1 3.29 2.22c1.92 3.29 5 2.34 6.26 1.79a4.61 4.61 0 0 1 1.37-2.88c-4.78-.54-9.8-2.38-9.8-10.62a8.29 8.29 0 0 1 2.22-5.77a7.68 7.68 0 0 1 .21-5.69s1.8-.58 5.91 2.2a20.46 20.46 0 0 1 10.76 0c4.11-2.78 5.91-2.2 5.91-2.2a7.74 7.74 0 0 1 .21 5.69a8.28 8.28 0 0 1 2.21 5.77c0 8.26-5 10.07-9.81 10.61a5.12 5.12 0 0 1 1.46 4v5.9c0 .71.39 1.24 1.48 1A21.5 21.5 0 0 0 24 2.5" />
       </svg>
-      <a [href]="project?.html_url" target="_blank">
-        <svg xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img" width="1em" height="1em"
-          preserveAspectRatio="xMidYMid meet" viewBox="0 0 48 48" width="1.5em" height="1.5em">
-          <path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round"
-            d="M24 2.5a21.5 21.5 0 0 0-6.8 41.9c1.08.2 1.47-.46 1.47-1v-3.65c-6 1.3-7.24-2.88-7.24-2.88A5.7 5.7 0 0 0 9 33.68c-1.95-1.33.15-1.31.15-1.31a4.52 4.52 0 0 1 3.29 2.22c1.92 3.29 5 2.34 6.26 1.79a4.61 4.61 0 0 1 1.37-2.88c-4.78-.54-9.8-2.38-9.8-10.62a8.29 8.29 0 0 1 2.22-5.77a7.68 7.68 0 0 1 .21-5.69s1.8-.58 5.91 2.2a20.46 20.46 0 0 1 10.76 0c4.11-2.78 5.91-2.2 5.91-2.2a7.74 7.74 0 0 1 .21 5.69a8.28 8.28 0 0 1 2.21 5.77c0 8.26-5 10.07-9.81 10.61a5.12 5.12 0 0 1 1.46 4v5.9c0 .71.39 1.24 1.48 1A21.5 21.5 0 0 0 24 2.5" />
-        </svg>
-      </a>
-      <a *ngIf="project?.homepage" [href]=" project?.homepage" target="_blank">
-        <svg xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img" width="1em" height="1em"
-          preserveAspectRatio="xMidYMid meet" viewBox="0 0 32 32" width="1.5em" height="1.5em">
-          <path fill="currentColor"
-            d="M30.94 15.66A16.69 16.69 0 0 0 16 5A16.69 16.69 0 0 0 1.06 15.66a1 1 0 0 0 0 .68A16.69 16.69 0 0 0 16 27a16.69 16.69 0 0 0 14.94-10.66a1 1 0 0 0 0-.68ZM16 25c-5.3 0-10.9-3.93-12.93-9C5.1 10.93 10.7 7 16 7s10.9 3.93 12.93 9C26.9 21.07 21.3 25 16 25Z" />
-          <path fill="currentColor" d="M16 10a6 6 0 1 0 6 6a6 6 0 0 0-6-6Zm0 10a4 4 0 1 1 4-4a4 4 0 0 1-4 4Z" /></svg>
-      </a>
-    </div>
-  </article>
-</section>
+    </a>
+    <a *ngIf="project?.homepage" [href]=" project?.homepage" target="_blank">
+      <svg xmlns="http://www.w3.org/2000/svg" aria-hidden="true" role="img" width="1em" height="1em"
+        preserveAspectRatio="xMidYMid meet" viewBox="0 0 32 32" width="1.5em" height="1.5em">
+        <path fill="currentColor"
+          d="M30.94 15.66A16.69 16.69 0 0 0 16 5A16.69 16.69 0 0 0 1.06 15.66a1 1 0 0 0 0 .68A16.69 16.69 0 0 0 16 27a16.69 16.69 0 0 0 14.94-10.66a1 1 0 0 0 0-.68ZM16 25c-5.3 0-10.9-3.93-12.93-9C5.1 10.93 10.7 7 16 7s10.9 3.93 12.93 9C26.9 21.07 21.3 25 16 25Z" />
+        <path fill="currentColor" d="M16 10a6 6 0 1 0 6 6a6 6 0 0 0-6-6Zm0 10a4 4 0 1 1 4-4a4 4 0 0 1-4 4Z" /></svg>
+    </a>
+  </div>
+</article>
