+++ src/app/components/projects/projects/projects.component.scss
@@ -1,49 +1,40 @@
-.projects {
+.cards {
     display: flex;
-    flex-direction: column;
+    flex-flow: row wrap;
+    justify-content: space-around;
     width: 100%;
     min-width: 1000px;
-    height: 100%;
-    gap: 1em;
-    margin: 16px 0;
+    border-radius: 5px;
+    border: 1px solid #e5e5e5;
 
-    &__cards {
-        display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
-        justify-content: space-between;
-        width: 100%;
-        height: 100%;
-        // border: 1px solid #e6e6e6;
+    &__info {
+        width: 90%;
+        padding: 1em;
 
-        &-info {
-            width: 90%;
-
-            p {
-                margin: 12px 0;
-                padding: 0;
-            }
-
-            h2 {
-                margin: 0 0 16px 0;
-                padding: 0;
-            }
+        p {
+            margin: 12px 0;
+            padding: 0;
         }
 
-        &-actions {
-            display: flex;
-            flex-direction: column;
-            gap: 0.75em;
+        h2 {
+            margin: 0 0 16px 0;
+            padding: 0;
+        }
+    }
+
+    &__actions {
+        display: flex;
+        flex-direction: column;
+        gap: 0.75em;
+        color: #333;
+        justify-content: center;
+
+        a {
             color: #333;
-            justify-content: center;
+            text-decoration: none;
 
-            a {
-                color: #333;
-                text-decoration: none;
-
-                :focus {
-                    color: #555;
-                }
+            :focus {
+                color: #555;
             }
         }
     }
