+++ src/app/components/videos/videos/videos.component.scss
@@ -1,36 +1,24 @@
-.videos {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: flex-start;
-    height: auto;
-    height: auto;
-    width: 100%;
-    min-width: 1000px;
-    gap: .5rem;
+.card {
+    width: 240px;
+    height: 350px;
 
-    &__card {
-        width: 240px;
-        height: 350px;
+    a {
+        color: #fff;
+        text-decoration: none;
+    }
 
-        a {
-            color: #fff;
-            text-decoration: none;
-        }
+    img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+    }
 
-        img {
-            width: 100%;
-            height: 100%;
-            object-fit: cover;
-        }
-
-        h4 {
-            position: relative;
-            margin: 0;
-            padding: 0;
-            top: -54px;
-            text-align: center;
-            color: #fff;
-        }
+    h4 {
+        position: relative;
+        margin: 0;
+        padding: 0;
+        top: -54px;
+        text-align: center;
+        color: #fff;
     }
 }
