+++ src/app/pages/portal/principal/principal.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { IPost } from '@core/models/devto/response-devto';
+import { IResponseGithub } from '@core/models/github/response-github';
 import { IResponseStatistics } from '@core/models/statistic/response-statistec';
 import { RequestYoutube } from '@core/models/youtube/request-youtube';
-import { ResponseYoutube } from '@core/models/youtube/response-youtube';
+import { Item, ResponseYoutube } from '@core/models/youtube/response-youtube';
 import { DevtoService } from '@core/services/devto/devto.service';
+import { GithubService } from '@core/services/github/github.service';
 import { YoutubeService } from '@core/services/youtube/youtube.service';
 
 @Component({
@@ -15,13 +17,18 @@ export class PrincipalComponent implements OnInit {
   datos = [1, 2, 3];
   dataYoutube!: ResponseYoutube;
   posts!: IPost[];
+  projects!: IResponseGithub[];
+  videos!: Item[];
   constructor(
     private youtubeService: YoutubeService,
-    private devtoService: DevtoService
+    private devtoService: DevtoService,
+    private githubService: GithubService
   ) {}
   ngOnInit(): void {
     this.getDataYoutube();
     this.getPostsDevto();
+    this.getAllProjects();
+    this.getAllVideos();
   }
 
   getDataYoutube(): void {
@@ -34,6 +41,19 @@ export class PrincipalComponent implements OnInit {
   getPostsDevto(): void {
     this.devtoService.getPostDevTo().subscribe((data: IPost[]) => {
       this.posts = data.slice(0, 3);
+    });
+  }
+
+  getAllProjects(): void {
+    this.githubService.getAllProjects().subscribe((data: IResponseGithub[]) => {
+      this.projects = data;
+    });
+  }
+
+  getAllVideos(): void {
+    this.youtubeService.getLastesYoutubeVideos().subscribe((data) => {
+      console.log(data);
+      this.videos = data.items;
     });
   }
 }
